Add the exert icon to repo
This icon shows up on the Elsa / Snow Queen card. We're not certain that
the name of it is "exert", but it looks like it represents what you'll
do to Elsa in order to do the same to another card. I've also referred
to this as "tap" in other code.