Add the pip icon and inkpot icon.
This "pip" icon is a mysterious little pip. I wish we knew more
about it!

"Inkpot" or perhaps otherwise known as the "cost" icon, is in the top
left corner of the card. Actions seemingly use the plain "cost" hexagon,
but characters have the Inkpot. It's unclear whether the distinction is
merely visual or not.